Finish the demo 2

+++ Project 1 - Shopping Bag/src/ShoppingBag.java
@@ -70,7 +70,7 @@ public class ShoppingBag {
 
     //find the total price of items in the shopping bag
     public double salesPrice() {
-        int total_sales = 0;
+        double total_sales = 0;
         for (int i = 0; i < size; i++) {
             total_sales += bag[i].price;
         }
